Added divs into my index:
Then added styles to to my nav to have an affect for when i hover over it. I also added a border under my nav when i hover to add a nice look to it. Lastly i added affects to each one of my divs and then added animation styles.